@@ -12,17 +12,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-material-components/datetime-picker": "^6.0.3",
-    "@angular/animations": "~12.2.0",
-    "@angular/cdk": "^12.2.4",
-    "@angular/common": "~12.2.0",
-    "@angular/compiler": "~12.2.0",
-    "@angular/core": "~12.2.0",
-    "@angular/forms": "~12.2.0",
-    "@angular/material": "^12.2.4",
-    "@angular/platform-browser": "~12.2.0",
-    "@angular/platform-browser-dynamic": "~12.2.0",
-    "@angular/router": "~12.2.0",
+    "@angular-material-components/datetime-picker": "https://github.com/MathewBerg/angular-material-components.git#develop_angular14",
+    "@angular/animations": "~14.0.4",
+    "@angular/cdk": "^14.0.4",
+    "@angular/common": "~14.0.4",
+    "@angular/compiler": "~14.0.4",
+    "@angular/core": "~14.0.4",
+    "@angular/forms": "~14.0.4",
+    "@angular/material": "^14.0.4",
+    "@angular/platform-browser": "~14.0.4",
+    "@angular/platform-browser-dynamic": "~14.0.4",
+    "@angular/router": "~14.0.4",
     "@juggle/resize-observer": "^3.3.1",
     "@ngx-translate/core": "^13.0.0",
     "@ngx-translate/http-loader": "^6.0.0",
@@ -42,9 +42,9 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.2.4",
-    "@angular/cli": "~12.2.4",
-    "@angular/compiler-cli": "~12.2.0",
+    "@angular-devkit/build-angular": "~14.0.4",
+    "@angular/cli": "~14.0.4",
+    "@angular/compiler-cli": "~14.0.4",
     "@types/jasmine": "~3.8.0",
     "@types/node": "^12.11.1",
     "jasmine-core": "~3.8.0",
@@ -54,6 +54,6 @@
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
     "ng-swagger-gen": "^2.3.1",
-    "typescript": "~4.3.5"
+    "typescript": "~4.7.4"
   }
 }
