@@ -4,9 +4,8 @@ import { SharedModule } from 'app/shared/shared.module';
 import { ConfirmDialogComponent } from './confirm-dialog.component';
 
 @NgModule({
-  imports: [SharedModule],
-  declarations: [ConfirmDialogComponent],
-  exports: [ConfirmDialogComponent],
-  entryComponents: [ConfirmDialogComponent]
+    imports: [SharedModule],
+    declarations: [ConfirmDialogComponent],
+    exports: [ConfirmDialogComponent]
 })
 export class ConfirmDialogModule {}
